+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Everything is released. Yay! :tada:
+
+## [1.8.0] - 2019-10-10
+
 - [Date.today(), Date.yesterday(), Date.tomorrow() (#9)](https://github.com/ngarbezza/oow/issues/9)
 - [day(), monthNumber(), year() for Date objects (#10)](https://github.com/ngarbezza/oow/issues/10)
 - [Point class with essential behavior (#11)](https://github.com/ngarbezza/oow/issues/11)
@@ -33,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changes to project structure: readme, changelog, dotfiles.
 - Upgrade to latest version of Testy (v3.1.1)
 
-[Unreleased]: https://github.com/ngarbezza/testy/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/ngarbezza/testy/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/ngarbezza/testy/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/ngarbezza/testy/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/ngarbezza/testy/compare/v1.5.1...v1.6.1
 [1.5.1]: https://github.com/ngarbezza/testy/compare/v1.5.0...v1.5.1